+++ main.py
@@ -1,7 +1,8 @@
 import torch
 from dynamics.dynamics import LinearDynamics
 from distributions.distributions import Gaussian
-from total_variation.experiments import approximation_scheme_tv
+from plotting.plotting import plot_samples
+from total_variation.experiments import monte_carlo, sample_from_gmm, propagate_mixture_tv_bounds
 from grid.regions import HyperRectangle, HyperRectangularPartition
 from grid.utils import get_shell_loc, uniform_grid
 
@@ -11,41 +12,40 @@ if __name__ == '__main__':
     # System parameters
     A = torch.Tensor(
         [
-            [0.9]
+            [0.9, 0.01],
+            [0.02, 0.5]
         ])
     f = LinearDynamics(A)
 
     # Initial distribution
-    mean_initial = torch.Tensor([2])
-    cov_initial = 0.5 * torch.eye(1)
+    mean_initial = torch.Tensor([2, 3])
+    cov_initial = 0.02 * torch.eye(2)
     initial_distribution = Gaussian(mean_initial, cov_initial)
 
     # Noise distribution
-    mean_noise = torch.Tensor([0])
-    cov_noise = 0.1 * torch.eye(1)
+    mean_noise = torch.Tensor([0, 0])
+    cov_noise = 0.1 * torch.eye(2)
     noise_distribution = Gaussian(mean_noise, cov_noise)
 
-    # Define partition
-    shell = torch.tensor([
-        [-8.],
-        [8.]
-    ])
-    loc_shell = get_shell_loc(shell)
-    inner_locs = uniform_grid(shell[0], shell[1], 100)
-
-    partition = HyperRectangularPartition(inner_locs, loc_shell, shell)
-
     # Define unsafe set
-    unsafe_set = HyperRectangle(torch.tensor([3.0]).unsqueeze(0), torch.tensor([5.0]).unsqueeze(0))
+    unsafe_set = HyperRectangle(torch.tensor([3.0, 4]).unsqueeze(0), torch.tensor([5.0, 6]).unsqueeze(0))
 
-    mixtures, tv_bounds = approximation_scheme_tv(f=f,
+    mixtures, tv_bounds = propagate_mixture_tv_bounds(f=f,
                                                   initial_distribution=initial_distribution,
                                                   noise_distribution=noise_distribution,
                                                   initial_grid_size=10,
-                                                  prediction_horizon=2,
+                                                  prediction_horizon=10,
                                                   n_samples=1000)
 
     print(tv_bounds)
 
-    # mixtures_hitting_probs = monte_carlo.mixture_approximation_monte_carlo(mixtures, barrier, parameters.n_samples)
-    # print(f"Mixtures hitting probs: {mixtures_hitting_probs}")
+    monte_carlo_samples = monte_carlo(f=f,
+                initial_distribution=initial_distribution,
+                noise_distribution=noise_distribution,
+                prediction_horizon=10,
+                n_samples=1000)
+
+    gmm_samples = sample_from_gmm(mixtures=mixtures,
+                    n_samples=1000)
+
+    plot_samples(monte_carlo_samples, gmm_samples)
