+++ plotting/plotting.py
@@ -19,25 +19,3 @@ def plot_samples(monte_carlo_samples: list,
 
     plt.tight_layout()
     plt.show()
-
-def plot_interval(gmm_prob_set: torch.Tensor,
-                  alpha_set: torch.Tensor,
-                  beta_set: torch.Tensor,
-                  actual_set_prob: torch.Tensor = None):
-
-    t = torch.arange(len(gmm_prob_set))
-    lbs = gmm_prob_set + alpha_set
-    ubs = gmm_prob_set + beta_set
-
-    plt.fill_between(t, lbs, ubs, color="lightgrey", label = r'$[\hat{\mathbb{P}}_{x_t}(U) + \alpha_{U, t}, \hat{\mathbb{P}}_{x_t}(U) + \beta_{U, t}]$')
-
-    plt.plot(t, gmm_prob_set, label=r'$\hat{\mathbb{P}}_{x_t}(U)$', linestyle='-', marker='s', color='red')
-    if actual_set_prob is not None:
-        plt.plot(t, actual_set_prob, label=r'$\mathbb{P}_{x_t}(U)$', linestyle='-', marker='s', color='green')
-
-    plt.xlabel("Time step")
-    plt.ylabel("Probability")
-    plt.legend(loc="upper right")
-    plt.grid(True)
-
-    plt.show()
